+++ templates/Businesses/index.php
@@ -17,7 +17,9 @@
     <div class="panel panel-default articles">
         <div class="panel-heading">
             Corporate Groups
+            <?php if($user_connected['role_id'] != 2 || $auths[38]) : ?>
             <a class="btn btn-warning" style="float:right" href="<?= ROOT_DIREC ?>/businesses/add">New</a>
+        <?php endif; ?>
         </div>
     <div class="panel-body articles-container">
         <div class="table-responsive">
@@ -26,8 +28,12 @@
                     <th>#</th>
                     <th class="text-center">Name</th>
                     <th class="text-center">Groups</th>
+                    <?php if($user_connected['role_id'] != 2 || $auths[37]) : ?>
                     <th class="text-center">Premium</th>
+                <?php endif; ?>
+                    <?php if($user_connected['role_id'] != 2 || $auths[38]) : ?>
                     <th class="text-center"></th>
+                <?php endif; ?>
                 </thead>
             <tbody> 
             <?php foreach($businesses as $business) : ?>
@@ -49,13 +55,17 @@
                     <td><a href="<?= ROOT_DIREC ?>/businesses/view/<?= $business->id ?>"><?= $business->business_number ?></a></td>
                     <td class="text-center"><?= $business->name ?></td>
                     <td class="text-center"><?= count($business->groupings) ?></td>
+                    <?php if($user_connected['role_id'] != 2 || $auths[37]) : ?>
                     <td class="text-center"><?= number_format($total, 2, ".", ",") ?></td>
+                <?php endif; ?>
+                <?php if($user_connected['role_id'] != 2 || $auths[38]) : ?>
                     <td class="text-right">
                         <a href="<?= ROOT_DIREC ?>/businesses/edit/<?= $business->id ?>" style="font-size:1.3em!important;"><span class="fa fa-xl fa-pencil color-blue"></span></a>
                         <?php if(empty($business->groupings) && empty($business->renewals)) : ?>
                         <a href="<?= ROOT_DIREC ?>/businesses/delete/<?= $business->id ?>" onclick="return confirm('Are you sure you would like to delete the company <?= $business->name ?>')" style="font-size:1.3em!important;margin-left:5px"><span class="fa fa-xl fa-trash color-red"></span></a>
                         <?php endif; ?>
                     </td>
+                <?php endif; ?>
                 </tr>
             <?php endforeach; ?>
             </tbody>
