+++ templates/Types/index.php
@@ -21,7 +21,9 @@
     <div class="panel panel-default articles">
         <div class="panel-heading">
             Settings
+            <?php if($user_connected['role_id'] != 2 || $auths[57]) : ?>
             <a class="btn btn-warning" style="float:right" href="<?= ROOT_DIREC ?>/types/add">New</a>
+            <?php endif; ?>
         </div>
     <div class="panel-body articles-container">
         <div class="table-responsive">
@@ -30,20 +32,27 @@
                     <th>Name</th>
                     <th class="text-center">Color</th>
                     <th class="text-center">Is Deductible</th>
+                    <?php if($user_connected['role_id'] != 2 || $auths[57]) : ?>
                     <th class="text-center"></th>
+                    <?php endif; ?>
                 </thead>
             <tbody> 
             <?php foreach($types as $type) : ?>
                 <tr>
                     <td><?= $type->name ?></td>
                     <td class="text-center"><div style="height:20px;width:20px;background-color:<?= $type->color ?>;margin:auto;border:1px solid black;border-radius:3px"></div></td>
-
+                    <?php if($user_connected['role_id'] != 2 || $auths[57]) : ?>
                     <td class="text-center"><input type="radio" value="<?= $type->id ?>" class="is_deductible" name="is_deductible" <?= ($type->is_deductible == 1) ? "checked" : "" ?>></td>
-                    <td class="text-right">
-                        <a href="<?= ROOT_DIREC ?>/types/edit/<?= $type->id ?>" style="font-size:1.3em!important;"><span class="fa fa-xl fa-pencil color-blue"></span></a>
-                        <?php if(empty($type->claims_types)) : ?>
-                        <a href="<?= ROOT_DIREC ?>/types/delete/<?= $type->id ?>" onclick="return confirm('Are you sure you would like to delete the type <?= $type->name ?>')" style="font-size:1.3em!important;margin-left:5px"><span class="fa fa-xl fa-trash color-red"></span></a>
-                    <?php   endif; ?>
+                    <?php else : ?>
+                        <td class="text-center"><?= ($type->is_deductible == 1) ? "Yes" : "" ?></td>
+                    <?php endif; ?>
+                    <?php if($user_connected['role_id'] != 2 || $auths[57]) : ?>
+                        <td class="text-right">
+                            <a href="<?= ROOT_DIREC ?>/types/edit/<?= $type->id ?>" style="font-size:1.3em!important;"><span class="fa fa-xl fa-pencil color-blue"></span></a>
+                            <?php if(empty($type->claims_types)) : ?>
+                            <a href="<?= ROOT_DIREC ?>/types/delete/<?= $type->id ?>" onclick="return confirm('Are you sure you would like to delete the type <?= $type->name ?>')" style="font-size:1.3em!important;margin-left:5px"><span class="fa fa-xl fa-trash color-red"></span></a>
+                            <?php   endif; ?>
+                    <?php endif; ?>
                     </td>
                 </tr>
             <?php endforeach; ?>
